Initial commit of database implementation
This commit contains editing and/or adding of 3 files.

DBDriver was edited, adding all its methods and comments for how they should be called.

database_database_initialization.sql contains the tables, procedures and functions for the database. It still needs the implementation of a few delete methods, but they will be added in a future implementation.

custom_data_for_database.sql contains some custom data for the database so the methods can be tested. The data is not intended for use in the final project, but only as a way for us to test the methods.