Initial commit of deletion procedures
This commit contains a change to the DBDriver.java and database_initialization.sql files.

The change for DBDriver is 5 new unsupported methods. These should be implemented later.

The change for the database_initialization file, are 5 new procedures for deleting different things in the database. There has also been made a change to the product table, where the reference it stores for a product category and a manufacture now can be null. This is required if you wish to delete a category or manufacture, because you need to set the value to null.