Extracted test method from DBDriverTest into smaller test classes
These classes now inherits from an super abstract DBDriverAbstractTest
class. This means that the testing of the very large and complex
DBDriver class is broken into smaller test classes, which makes it much
more easier to read and maintain an overview of.

Have a look at the header comment in DBDriverAbstractTest.java file