Fix fatal error in productByUUIDExists method
Fixes a fatal error which would render the connection to the database in
broken state.
If a malformed UUID string was passed in to the method, the SQL
connection would be left in broken state, because the thrown exception
was silently treated as being false. The broken state would lead to
unexpected thrown exception in subsequent method calls. This led to
hard to debug exceptions, and very unexpected results.