Signature seal of commit tree
This project has been developed on an internal Gitlab system,
which has required the use of private emails and names. Furthermore
some developers have chosen to remain completely anonymous. This means
that this repository has been modified to change committer and author
identities, to protect developers private identities.
This has been done using the https://github.com/newren/git-filter-repo
tool.
As a result of this alteration, all previous PGP signatures of commits
has been lost.
Through the PGP signing of this commit, I MidnightRocket declare that I
have NOT altered this repository, besides those modifications needed to
alter committer and author identities.

Provided that the git commit history is immutable, and that the
git-filter-repo tool can be trusted, all previous commits in the
ancestry path of this commit, attributed to me MidnightRocket can be
trusted as being authored and committed by me.

I cannot make any claims on the trustworthiness of any commits
attributed to any other contributor. I can only attest that I have left
the file tree of each commit as is without modification.